feat(assert): assert handler and led driver
Create an assert handler suitable for a microcontroller. Let it trigger a software breakpoint
when a debugger is attached and then enter an endless while loop blinking the test led.
The handler will later be extended with console tracking and turning off the motors.

Also, implement a simple led driver to avoid having to interact with the IO functions directly,
but don't use this driver in the assert handler to avoid it from being called recursively.

Lastly, wrap the __delay_cycles intrinsic in a new define macro BUSY_WAIT_ms that takes
milliseconds (instead of cycles) as argument.