build: add two custom target in cmake for analyzing flash sace used
msp430g2553 only has 16KB of FLASH, so care must be taken when programming it as to not run out
of this small space. The GCC-toolchain has several tools for seeing how much space the code
occupies, add these as rules to the Makefile.

While at it, remove the inline from the IO functions because they apparently occupy more space as
inlined.