fix(init): increase clock speed from  1 MHz to 16 MHz
increase the clock speed to 16 MHz mostly because I can. The clock speed is not the limiting factor
in this application so technically there is no reason to run it at 16 MHz, but the power
consumption is negligible compared to other things, and if I'm going to change it it's better
to do it now rather than later.