feat(drivers): UART driver and ring buffer
Add a driver for the UART peripheral that can transmit individual characters. Make it interrupt
driven and put the characters in a ring buffer that the interrupt can feed from. This is going
to be used to implement tracing (printf) in the future.

Also, write the calibrated clock values correctly.

This was tested by connecting the LAUNCHPAD to a PC via a USB-to-UART bridge and using the command:
picocom -b 115200 /dev/ttyUSB0