feat(trace): use external printf implementation for trace
The standard library printf implementation is not appropriate for a microcontroller.
A microcontroller has no notion of stdout and the standard printf implementation has a big memory
footprint. To get around this, use a more stripped-down external implementation, mpaland/printf,
and include it as a git submodule (already included). Use this to implement a trace function that
can format strings.